# frontend/openapi.yaml
openapi: 3.1.0
info:
  title: Sarathi Next API
  version: 1.0.0
  description: Contract for the Sarathi Next Learner's Licence prototype integration.
servers:
  - url: https://api.example.com
    description: Replace with the deployed API origin.
...
paths:
...
  /v1/auth/mobile/challenges:
    post:
      tags: [Authentication]
      operationId: createMobileChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobileNumber, locale]
              properties:
                mobileNumber: { type: string, pattern: '^[0-9]{10}$', example: '9876543210' }
                locale: { $ref: '#/components/schemas/Locale' }
      responses:
        '201':
          description: OTP challenge created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallenge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/mobile/challenges/{challengeId}/verify:
    post:
      tags: [Authentication]
      operationId: verifyMobileChallenge
      parameters:
        - $ref: '#/components/parameters/ChallengeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp, consentToStatusUpdates]
              properties:
                otp: { type: string, pattern: '^[0-9]{6}$' }
                consentToStatusUpdates: { type: boolean, const: true }
      responses:
        '200':
          description: Verification succeeded and session cookie is set.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/auth/aadhaar/challenges:
    post:
      tags: [Authentication]
      operationId: createAadhaarChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aadhaarNumber, locale]
              properties:
                aadhaarNumber: { type: string, pattern: '^[0-9]{12}$' }
                locale: { $ref: '#/components/schemas/Locale' }
      responses:
        '201':
          description: Aadhaar OTP challenge created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallenge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/aadhaar/challenges/{challengeId}/verify:
    post:
      tags: [Authentication]
      operationId: verifyAadhaarChallenge
      parameters:
        - $ref: '#/components/parameters/ChallengeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp, declarationsAccepted]
              properties:
                otp: { type: string, pattern: '^[0-9]{6}$' }
                declarationsAccepted:
                  type: array
                  minItems: 3
                  maxItems: 3
                  items: { type: string, enum: [identity_verification, transaction_only, confidentiality] }
      responses:
        '200':
          description: Verification succeeded and session cookie is set.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /v1/applications/{applicationId}/tests:
    post:
      tags: [Tests]
      operationId: startTest
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      responses:
        '201':
          description: Test attempt started.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestAttempt' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tests/{testId}/answers:
    post:
      tags: [Tests]
      operationId: submitTestAnswer
      security: [{ sessionCookie: [] }]
      parameters:
        - name: testId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionId, answerId]
              properties:
                questionId: { type: string }
                answerId: { type: string }
      responses:
        '204': { description: Answer saved. }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/tests/{testId}/submit:
    post:
      tags: [Tests]
      operationId: submitTest
      security: [{ sessionCookie: [] }]
      parameters:
        - name: testId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Test scored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestResult' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sarathi_session
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, pattern: '^app_[A-Za-z0-9]+$' }
    ChallengeId:
      name: challengeId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    Locale: { type: string, enum: [en, hi] }
...
    OtpChallenge:
      type: object
      required: [challengeId, expiresAt, resendAfterSeconds, maskedDestination]
      properties:
        challengeId: { type: string }
        expiresAt: { type: string, format: date-time }
        resendAfterSeconds: { type: integer, minimum: 0 }
        maskedDestination: { type: string, example: '******3210' }
    AuthResult:
      type: object
      required: [authenticated, applicationId, nextStep]
      properties:
        authenticated: { type: boolean, const: true }
        applicationId: { type: string }
        nextStep: { type: string, enum: [personal_details, task_list] }
...
    TestQuestion:
      type: object
      required: [questionId, text, answers]
      properties:
        questionId: { type: string }
        text: { type: string }
        answers:
          type: array
          minItems: 2
          items:
            type: object
            required: [answerId, text]
            properties:
              answerId: { type: string }
              text: { type: string }
    TestAttempt:
      type: object
      required: [testId, questionCount, passMark, questions, expiresAt]
      properties:
        testId: { type: string }
        questionCount: { type: integer }
        passMark: { type: integer }
        questions: { type: array, items: { $ref: '#/components/schemas/TestQuestion' } }
        expiresAt: { type: string, format: date-time }
    TestResult:
      type: object
      required: [testId, score, total, passed, completedAt]
      properties:
        testId: { type: string }
        score: { type: integer }
        total: { type: integer }
        passed: { type: boolean }
        completedAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors:
              type: object
              additionalProperties: { type: string }
            requestId: { type: string }
  responses:
    BadRequest:
      description: Invalid request.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthenticated:
      description: Authentication required or OTP invalid.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    Conflict:
      description: Resource state conflict.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    RateLimited:
      description: Too many requests.
      headers: { Retry-After: { schema: { type: integer } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...